Rework the TreeBrush Command + Fix ArrayOutOfBoundsException issue
The tree generation didn't work properly. In order to generate a tree, we need to spawn it from a block of air, on top a valid block. I reworked the entire brush to handle a safe generation with the arrow, and a forced generation with the gunpowder. The safe generation lets Minecraft decide if the terrain is valid to spawn a tree. The forced one tries to create the conditions needed to generate with some obvious limitations.